Add generated clients for capi Machines
In order to use the CLI to get the SSH keys for machines, clients need
to be generated. This change sets up the clients to be generated with
`go generate`.

In addition to this change, the `config.json` file from the machine
state is also included in the response to the get request for SSH keys.
This is necessary because the SSH user information is stored there and
needed to connect.